Fixes #1
The issue is that there are some Blueprint buffs that the game doesn't like us disabling the effects for.  If the user disables some of these buffs (which don't originate from the player casting spells) and enters Drezen, then the game throws a critical error.

The correction I've taken is to be more selective in terms of the buffs you can tweak in the Buffs.json file.  I've limited buffs to mostly those which players trigger through spells or abilities, so ~950 buffs.  Testing no longer sees the problem repro.